Add files via upload
Think Python, Exercise 1_2
Q1. How many seconds are there in 42 minutes and 42 seconds?
Q2. How many miles are there in 10 kilometers?
Q3. If you run a 10 kilometer race in 42 minutes 42 seconds, what is your average pace (time per mile in minutes and seconds)?
      What is the average speed in miles per hour?
These solutions are presented for the user to input the time and distance variables, which makes it adaptable for any race.